Convert units of measure in a fact dataset

This function converts the units of measure in a fact dataset using a dataset of factors of conversions between units. The data frames of fact and of factors of conversions must be properly structured. For structures of data frames, see details and here: http://.

convert_units(df_input, df_conversion_factor, codelist_geoidentifiers_conversion_factors)

Arguments

df_input
a data.frame of fact

df_conversion_factor
a data.frame of factors of conversion between units

codelist_geoidentifiers_conversion_factors
NULL or string. The name of the coding system for the spatial dimension used in df_conversion_factor (i.e. table name in Sardara database), or NULL if the coding system for the spatial dimension is the same as the one used in df_input

Details

The data frame of factors of conversion between units (df_conversion_factor) must have the following structure: - several columns of dimension, that stratify the factor of conversion (e.g. species, gear, etc.), - one column named 'unit', that provides the code of the source unit of the conversion factor, - one column named 'unit_target', that provides the code of the target unit of the conversion factor, - one column named 'conversion_factor', that provides the numerical factor of conversion to convert the measure from the unit stated in the column 'unit' to the unit stated in the column 'unit_target'. See an example of dataset of factors of conversion here: https://goo.gl/KriwxV.

Example: These are the first rows of a dataset of factors of conversions:

source_authority 	 species 	 gear 	 geographic_identifier 	 time_start 	 time_end 	 unit 	 unit_target 	 conversion_factor

IOTC 	 YFT 	 LL 	 1 	 1952-01-01 	 1953-01-01 	 NO 	 MT 	 0.048060001

IOTC 	 YFT 	 LL 	 2 	 1952-01-01 	 1953-01-01 	 NO 	 MT 	 0.048680000

IOTC 	 YFT 	 LL 	 3 	 1952-01-01 	 1953-01-01 	 NO 	 MT 	 0.058639999

The first row means that for the combination of dimensions: source_authority=IOTC, species=YFT, gear=LL, geographical_identifier=1, starting date of validity of the factor of conversion (time_start)=1952-01-01, ending date of validity of the factor of conversion (time_start)=1953-01-01, the factor of conversion to convert a measure from unit=NO to target_unit=MT is equal to 0.048060001

The codes used in the dimensions of the dataset of factors of conversion (df_conversion_factor) must be the same as the ones used in the dataset of fact with units to convert (df_input), except for the spatial dimension (geographic_identifier) - see here-after for more details. The only mandatory columns of the dataset of factors of conversion are "unit", "unit_target" and "conversion_factor". All the other columns are here to stratify the factors of conversion (by species, gear, time, space, etc.). Particularly, the columns "time_start", "time_end" and "geographic_identifier" allow to stratify spatialy and temporarily the factors of conversion.

The columns "time_start" and "time_end" provide respectively the starting date and the ending date of validity of the factor of conversion.

The column "geographic_identifier" provides the spatial stratification of the factor of conversion. If the coding system for spatial stratification used in df_conversion_factor is the same as the one used in df_input, then the parameter codelist_geoidentifiers_conversion_factors must be set to NULL. Else, the spatial coding system used in df_conversion_factor must be stored in the Sardara database, and the parameter codelist_geoidentifiers_conversion_factors must be set to the name of the spatial coding system (table) in Sardara DB.

If df_conversion_factor mixes factors of conversion that have and do not have a spatial stratification, the rows that do not have spatial stratification must be set to geographic_identifier= 0.

Columns of time (time_start and time_end) must be of type character (not Posix) and they must have the same resolution (e.g. day, second. etc).

See also

Other create your own tuna atlas: map_codelist, raise_datasets_by_dimension, raise_get_rf

Examples

# Connect to Sardara World DB
con=db_connection_sardara_world()

#> Loading required package: DBI

# Retrieve IOTC georeferenced catch data from 2017
df_input<-iotc_catch_level0(2017)

#> Error in eval(expr, envir, enclos): impossible de trouver la fonction "iotc_catch_level0"

# some curation before use of the functions
df_input$time_start<-substr(as.character(df_input$time_start), 1, 10)

#> Error in substr(as.character(df_input$time_start), 1, 10): objet 'df_input' introuvable

df_input$time_end<-substr(as.character(df_input$time_end), 1, 10)

#> Error in substr(as.character(df_input$time_end), 1, 10): objet 'df_input' introuvable

colnames(df_input)[colnames(df_input) == "catchunit"] <- "unit"

#> Error in colnames(df_input)[colnames(df_input) == "catchunit"] <- "unit": objet 'df_input' introuvable

# Open a dataset of factors of conversion (the one used to convert units of catch in the IRD Tuna Atlas)
conversion_factors_dataset="https://goo.gl/KriwxV"
df_conversion_factor=read.csv(conversion_factors_dataset,stringsAsFactors = F,colClasses="character")
head(df_conversion_factor)

#>   source_authority species gear geographic_identifier time_start   time_end
#> 1             IOTC     YFT   LL                     1 1952-01-01 1953-01-01
#> 2             IOTC     YFT   LL                     2 1952-01-01 1953-01-01
#> 3             IOTC     YFT   LL                     3 1952-01-01 1953-01-01
#> 4             IOTC     BET   LL                     1 1952-01-01 1953-01-01
#> 5             IOTC     BET   LL                     2 1952-01-01 1953-01-01
#> 6             IOTC     BET   LL                     3 1952-01-01 1953-01-01
#>   unit unit_target  conversion_factor
#> 1   NO          MT  0.048060001373291
#> 2   NO          MT 0.0486800003051758
#> 3   NO          MT 0.0586399993896484
#> 4   NO          MT 0.0443400001525879
#> 5   NO          MT 0.0443400001525879
#> 6   NO          MT 0.0433499984741211

# Map gears of input dataset with standard gears, and then with sardara groups of gears, to align with codes of gears used in the conversion factor dataset
gear_rfmos_mapping_to_standard<-extract_dataset(con,list_metadata_datasets(con,dataset_name="codelist_mapping_gear_iotc_isscfg_revision_1"))

#> Error in nrow(metadata_dataset): impossible de trouver la fonction "list_metadata_datasets"

df_mapped<-map_codelist(df_input,gear_rfmos_mapping_to_standard,"gear")$df

#> 
#>  mapping dimension gear with code list mapping

#> Error in is.data.frame(x): objet 'df_input' introuvable

gear_standard_mapping_to_gear_sardara<-extract_dataset(con,list_metadata_datasets(con,dataset_name="codelist_mapping_gear_iotc_geargroup_tunaatlas"))

#> Error in nrow(metadata_dataset): impossible de trouver la fonction "list_metadata_datasets"

df_input<-map_codelist(df_input,gear_standard_mapping_to_gear_sardara,"gear")$df

#> 
#>  mapping dimension gear with code list mapping

#> Error in is.data.frame(x): objet 'df_input' introuvable

# Convert units MTNO to MT and remove NOMT (we do not keep the data that were expressed in number with corresponding value in weight)
df_input$unit[which(df_input$unit == "MTNO")]<-"MT"

#> Error in df_input$unit[which(df_input$unit == "MTNO")] <- "MT": objet 'df_input' introuvable

df_input<-df_input[!(df_input$unit=="NOMT"),]

#> Error in eval(expr, envir, enclos): objet 'df_input' introuvable

# Convert units from numbers to weight using the dataset of factors of conversion. 
# The spatial coding system used in conversion_factor (column geographic_identifier) is not the same as the one used in df_input. Hence, we set in the parameter codelist_geoidentifiers_conversion_factors the name of the spatial coding system used in df_conversion factor ("areas_conversion_factors_numtoweigth_ird").
df_converted<-convert_units(con, df_input = df_input, df_conversion_factor = df_conversion_factor, codelist_geoidentifiers_conversion_factors = "areas_conversion_factors_numtoweigth_ird")

#> 
#>  converting units and measures

#> Error in data.table(df_input): objet 'df_input' introuvable

# Get the dataframe with units converted: data that were expressed in number are converted to metric tons. Some data might not be converted at all because no conversion factor exists for the stratum: these data are kept in their original unit (in this case, number).
df_converted_df<-df_converted$df

#> Error in eval(expr, envir, enclos): objet 'df_converted' introuvable

head(df_converted_df)

#> Error in head(df_converted_df): objet 'df_converted_df' introuvable

# Get information regarding the conversion (data converted, data not converted because no factor of conversion existed, etc.)
df_converted$stats

#> Error in eval(expr, envir, enclos): objet 'df_converted' introuvable

dbDisconnect(con)

#> [1] TRUE