Extract a dataset available in Sardara database

This function outputs a dataset available in the Sardara database.

extract_dataset(con,metadata_dataset)

Arguments

con
a wrapper of rpostgresql connection (connection to a database)

metadata_dataset
data.frame of type "metadata" (one row extracted from the table metadata.metadata).

Value

a data.frame of the data available in the database and set as dataset_name

Details

The meaning of the columns of a raw dataset is provided in the data dictionary of raw datasets. ## to fill in this description when it is finalized

Examples

con=db_connection_sardara_world()

#> Error in postgresqlNewConnection(drv, ...): RS-DBI driver: (cannot allocate a new connection -- maximum of 16 connections already opened)

# Extract a raw dataset
raw_dataset<-extract_dataset(con,list_metadata_datasets(con,dataset_name="global_catch_5deg_1m_1950_01_01_2016_01_01_tunaatlasIRD_level1"))

#> Error in nrow(metadata_dataset): impossible de trouver la fonction "list_metadata_datasets"

# Extract a code list
code_list<-extract_dataset(con,list_metadata_datasets(con,dataset_name="species_asfis"))

#> Error in nrow(metadata_dataset): impossible de trouver la fonction "list_metadata_datasets"

# Extract a mapping between code lists )
code_list_mapping<-extract_dataset(con,list_metadata_datasets(con,dataset_name="codelist_mapping_species_iotc_speciesgroup_tunaatlas"))

#> Error in nrow(metadata_dataset): impossible de trouver la fonction "list_metadata_datasets"