Get a list of queries to execute to plug services to the datasets available in Sardara

Get a list of queries to execute to manage the datasets available in Sardara (extract in csv format, in netcdf, etc.)

getSQLSardaraQueries(con,dataset_metadata)

Arguments

con
a wrapper of rpostgresql connection (connection to a database)

dataset_metadata
data.frame of type "metadata" (one row extracted from the table metadata.metadata).

Value

a list with SQL queries and informations on the dataset. Depending on the type of data (raw_dataset, codelist, mapping), the output elements of the list might vary. The elements of the list are:

query_CSV : Query to retrieve a data.frame

query_NetCDF :

query_wfs_wms :

query_wfs_wms_aggregated_layer :

query_dynamic_list_keywords_species :

query_dynamic_list_keywords_fishing_gears :

query_dynamic_list_keywords_Fleets :

query_dynamic_metadata_count_features :

query_dynamic_metadata_spatial_Extent :

query_dynamic_metadata_temporal_Extent :

query_dynamic_metadata_get_SRID :

lineage :

query_dynamic_list_keywords_institutions :

Examples

 con=db_connection_sardara_world()

#> Error in postgresqlNewConnection(drv, ...): RS-DBI driver: (cannot allocate a new connection -- maximum of 16 connections already opened)

 # retrieve metadata row of dataset global_catch_5deg_1m_1950_01_01_2016_01_01_tunaatlasIRD_level1
dataset_metadata<-list_metadata_datasets(con,dataset_name="global_catch_5deg_1m_1950_01_01_2016_01_01_tunaatlasIRD_level1")

#> Error in eval(expr, envir, enclos): impossible de trouver la fonction "list_metadata_datasets"

queries<-getSQLSardaraQueries(con,dataset_metadata)

#> Error in postgresqlQuickSQL(conn, statement, ...): expired PostgreSQLConnection

# retrieve data.frame of global_catch_5deg_1m_1950_01_01_2016_01_01_tunaatlasIRD_level1

global_catch_5deg_1m_1950_01_01_2016_01_01_tunaatlasIRD_level1<-dbGetQuery(con,queries$query_CSV)

#> Error in dbGetQuery(con, queries$query_CSV): objet 'queries' introuvable