Extract source georeferenced efforts datasets of IOTC from Sardara World database

This function extracts the source georeferenced effort datasets stored in the Sardara World database coming from the Indian Ocean Tuna Commission (IOTC). The output dataset provides efforts of tuna, tuna-like and shark species in the Indian ocean. Efforts are stratified by month, gear, vessel flag reporting country, fishing mode (i.e. type of school used), area (usualy 1° or 5° square) and unit of effort. Data are expressed using IOTC's coding system.

iotc_effort_level0(year_tunaatlas)

Arguments

year_tunaatlas
numeric. The year of the datasets to extract. Starts in 2017

Details

Examples

# Retrieve IOTC georeferenced efforts data from 2017
iotc_effort<-iotc_effort_level0(2017)

#> Error in eval(expr, envir, enclos): impossible de trouver la fonction "iotc_effort_level0"

head(iotc_effort)

#> Error in head(iotc_effort): objet 'iotc_effort' introuvable