List metadata of the datasets available in Sardara database

This function outputs the metadata of the data available in the Sardara database.

list_metadata_datasets(con,dataset_name=NULL,source_authority=NULL)
list_metadata_raw_datasets(con,dataset_name=NULL,source_authority=NULL,variable=NULL,spatial_resolution=NULL,level_of_correction=NULL)
list_metadata_codelists(con,dataset_name=NULL,source_authority=NULL,dimension=NULL)
list_metadata_codelists_mapping(con,dataset_name=NULL,source_authority=NULL,dimension=NULL)

Arguments

con
a wrapper of rpostgresql connection (connection to a database)

dataset_name
NULL or string. If not NULL, extracts the metadata row of the dataset_name stated. In this case, all the other parameters will be ignored

source_authority
NULL or vector of strings. If not NULL, filter available datasets by the source authority in charge of producing the source statistics collated and harmonized. E.g. c("IOTC","ICCAT") will provide the metadata only for the data produced by IOTC and ICCAT.

dimension
NULL or vector of strings. For codelists and mappings only. If not NULL, filter available code lists / mappings by dimensions. E.g. c("gear","species") will provide the metadata only for code lists/mappings between code lists related to fishing gears and species.

variable
NULL or vector of strings. For datasets only. If not NULL, filter available code lists / mappings by variable. Three variables are available in Sardara: catch, effort, catch_at_size

spatial_resolution
NULL or real. For datasets only. If not NULL, filter available datasets that are defined on that spatial resolution (in degrees).

level_of_correction
NULL or integer. For datasets only. If not NULL, filter available datasets that are have that level of correction provided.

Value

a data.frame of metadata for the data available in Sardara database. The meaning of the columns of the output data.frame is provided in the data dictionary of the metadata table of Sardara database. ## to fill in this description when it is finalized

Details

Three types of data are available in Sardara database:

"raw_dataset": Datasets. Three types of datasets are available:

"catch": Quantity of of fish in number or biomass harvested in a given stratum

"effort": Quantity of effort (expressed in a given unit such as number of sets, number of fishing hours, number of hooks, etc.) exerted in a given stratum

"catch-at-size": Quantity of fish at a given size (or size class) harvested in a given stratum

"codelist": Reference tables with at least codes, and usualy labels associated to these codes.

"mapping": Tables that establishes mappings / correspondences between codes from various reference tables.

The meaning of the columns of the raw dataset is provided in the data dictionary of raw datasets. ## to fill in this description when it is finalized

list_metadata_datasets lists the metadata of all the types of datasets (raw_dataset, codelists, mappings)

list_metadata_raw_datasets lists only the metadata of raw_datasets

list_metadata_codelists lists only the metadata of the code lists

list_metadata_mappings lists only the metadata of the mappings between code lists

Examples

# List the available source IOTC datasets:
metadata_iotc_datasets<-list_metadata_datasets(db_connection_sardara_world(),source_authority=c("IOTC"))

#> Error in eval(expr, envir, enclos): impossible de trouver la fonction "list_metadata_datasets"

# List the available code lists for WCPFC and IATTC
metadata_iccat_code_lists<-list_metadata_codelists(db_connection_sardara_world(),source_authority=c("WCPFC","IATTC"))

#> Error in eval(expr, envir, enclos): impossible de trouver la fonction "list_metadata_codelists"

# List the available raw datasets of catch and of effort that are defined on 5° grid resolution
metadata_raw_dataset_catch_5deg<-list_metadata_raw_datasets(db_connection_sardara_world(),variable=c("catch","effort"),spatial_resolution=5)

#> Error in eval(expr, envir, enclos): impossible de trouver la fonction "list_metadata_raw_datasets"