Maps and replaces one dimension of a fact dataset using a dataset of mappings between code lists

This function maps one dimension (i.e. column with codes) of a fact dataset to a target code list using a dataset of mappings between code lists. In other words, it makes the correspondance between two code lists for a given dimension, and replaces for this dimension in the fact dataset the old codes by the new codes available in the dataset of mappings between code lists.

map_codelist(df_input, df_mapping, dimension_to_map)

Arguments

df_input
a data.frame of fact

df_mapping
a data.frame of code list mapping

dimension_to_map
the name (string) of the dimension to map.

Value

a list with two objects:

"df": The input data.frame of fact, where the dimension_to_map has been mapped using the df_mapping

"stats": A data.frame with some information regarding the data not mapped. It provides, for each unit of measure available in the input dataset, the sum and percentage of the data that could not be map because no correspondance are available in the dataset of mappings between code lists

Details

The data frames of fact and code list mapping must be properly structured. The data.frame of mapping must at least have the 2 following columns:

"src_code": The source codes for the dimension to map (i.e. the codes used in the df_input fot the considered dimension)

"src_target": The target codes for the dimension to map

See also

Other create your own tuna atlas: convert_units, raise_datasets_by_dimension, raise_get_rf

Examples

  # Reads IOTC nominal catch dataset (2017 release)
  iotc_nominal_catch<-extract_dataset(con,list_metadata_datasets(con,dataset_name="indian_ocean_nominal_catch_1950_01_01_2015_01_01_tunaatlasIOTC_2017_level0"))

#> Error in nrow(metadata_dataset): impossible de trouver la fonction "list_metadata_datasets"

  head(iotc_nominal_catch)

#> Error in head(iotc_nominal_catch): objet 'iotc_nominal_catch' introuvable

  # Read a mapping between code lists (in this case, mapping between codes for fishing gears used by the tuna RFMOs and the International Standard Statistical Classification of Fishing Gear)
  df_mapping<-extract_dataset(con,list_metadata_datasets(con,dataset_name="codelist_mapping_gear_iotc_isscfg_revision_1"))

#> Error in nrow(metadata_dataset): impossible de trouver la fonction "list_metadata_datasets"

  head(df_mapping)

#> Error in head(df_mapping): objet 'df_mapping' introuvable

  # Map code lists. Output is a list with two elements (see section "return")
  df_mapped<-map_codelist(df_input,df_mapping,"gear")

#> 
#>  mapping dimension gear with code list mapping

#> Error in is.data.frame(x): objet 'df_input' introuvable

  # Get the dataframe mapped: dimension "gear" mapped to ISSCFG. The column "gear" has its values changed compared to the ones before the execution of the function. The codes have been mapped following the dimensions "gear" and "source_authority", since the dataset of mappings between code lists had both dimensions.
  df_mapped_df<-df_mapped$df

#> Error in eval(expr, envir, enclos): objet 'df_mapped' introuvable

  head(df_mapped_df)

#> Error in head(df_mapped_df): objet 'df_mapped_df' introuvable

  # Get information regarding the data that were not mapped.
  df_mapped$stats

#> Error in eval(expr, envir, enclos): objet 'df_mapped' introuvable