Create a dataset with most complete information using two datasets with partial information

This function creates a dataset with most complete information using two datasets with partial information , by crossing the informations available in both datasets. It takes as input two datasets, each with partial information, but the full information is available if we consider both datasets. Each dataset used as input misses one dimension, the missing dimension in the dataset n°1 being available in the dataset n°2 and reversely. The functions uses both datasets to create a new dataset with the full information (i.e. with values filled for both dimensions on the dataset). The data frames of fact must be properly structured. For structures of data frames, see details and here: http://.

raise_datasets_by_dimension(df1, df2, dimension_missing_df1, dimension_missing_df2)

Arguments

df1
a data.frame of fact with one dimension missing (i.e. for this dimension all values are set to UNK), the latter being available in df2.

df2
the same data.frame of fact with another dimension missing, the latter being available in df1.

dimension_missing_df1
string. The name of the dimension missing in df1.

dimension_missing_df2
string. The name of the dimension missing in df2.

Details

In this function, we make the hypothesis that ....? comment décrire ca???? -> cf mail Manu "Formes" Example:

df1 : is a dataset where the "flag" dimension is available but the "schooltype" dimension is missing (i.e. in this dataset, all the values of the column "schooltype" are set to UNK)

df2 : is a dataset where the "schooltype" dimension is available but the "flag" dimension is missing (i.e. in this dataset, all the values of the column "flag" are set to UNK)

The concept of raising the data is the following: in a given stratum S, a fishing country (flag) F has catched a percentage RF

In the output dataset, both flag and school type dimensions are available for each stratum.

We make the hypothesis that the proportion of catches by schooltype is equal for all the fishing countries.

See also

Other create your own tuna atlas: convert_units, map_codelist, raise_get_rf, spatial_curation_intersect_areas

Examples

## Retrieve some IATTC georeferenced times series of catch 
# Connect to Sardara DB
con <- db_connection_sardara_world()

# IATTC dataset stratified by schooltype (and not flag)
dataset_iattc_ce_PSSetType<-extract_dataset(con,list_metadata_datasets(con,dataset_name="east_pacific_ocean_catch_1958_12_01_2016_01_01_1deg_1m_ps_tunaatlasIATTC_2017_level0__tuna_bySchool"))

#> Error in nrow(metadata_dataset): impossible de trouver la fonction "list_metadata_datasets"

head(dataset_iattc_ce_PSSetType)

#> Error in head(dataset_iattc_ce_PSSetType): objet 'dataset_iattc_ce_PSSetType' introuvable

unique(dataset_iattc_ce_PSSetType$flag) # Note that the column "flag" is all set with "UNK"

#> Error in unique(dataset_iattc_ce_PSSetType$flag): objet 'dataset_iattc_ce_PSSetType' introuvable

unique(dataset_iattc_ce_PSSetType$schooltype) # Note that the column "schooltype" is detailed

#> Error in unique(dataset_iattc_ce_PSSetType$schooltype): objet 'dataset_iattc_ce_PSSetType' introuvable

# Same IATTC datasets, but stratified by flag (and not schooltype)
dataset_iattc_ce_PSFlag<-extract_dataset(con,list_metadata_datasets(con,dataset_name="east_pacific_ocean_catch_1958_12_01_2016_01_01_1deg_1m_ps_tunaatlasIATTC_2017_level0__tuna_byFlag"))

#> Error in nrow(metadata_dataset): impossible de trouver la fonction "list_metadata_datasets"

head(dataset_iattc_ce_PSFlag)

#> Error in head(dataset_iattc_ce_PSFlag): objet 'dataset_iattc_ce_PSFlag' introuvable

unique(dataset_iattc_ce_PSFlag$schooltype) # Note that the column "schooltype" is all set with "UNK"

#> Error in unique(dataset_iattc_ce_PSFlag$schooltype): objet 'dataset_iattc_ce_PSFlag' introuvable

unique(dataset_iattc_ce_PSFlag$flag) # Note that the column "flag" is detailed

#> Error in unique(dataset_iattc_ce_PSFlag$flag): objet 'dataset_iattc_ce_PSFlag' introuvable

## Raise both datasets. In the output dataset, both flag and school type information are available for each stratum. 

dataset_iattc_flag_raised_to_schooltype<-raise_datasets_by_dimension(
df1=dataset_iattc_ce_PSFlag,
df2=dataset_iattc_ce_PSSetType,
dimension_missing_df1="schooltype",
dimension_missing_df2="flag")

#> Error in eval(expr, envir, enclos): impossible de trouver la fonction "raise_datasets_by_dimension"

head(dataset_iattc_flag_raised_to_schooltype$df)

#> Error in head(dataset_iattc_flag_raised_to_schooltype$df): objet 'dataset_iattc_flag_raised_to_schooltype' introuvable

unique(dataset_iattc_flag_raised_to_schooltype$df$schooltype)

#> Error in unique(dataset_iattc_flag_raised_to_schooltype$df$schooltype): objet 'dataset_iattc_flag_raised_to_schooltype' introuvable

unique(dataset_iattc_flag_raised_to_schooltype$df$flag) # Note that both columns "flag" and "schooltype" are detailed

#> Error in unique(dataset_iattc_flag_raised_to_schooltype$df$flag): objet 'dataset_iattc_flag_raised_to_schooltype' introuvable

dbDisconnect(con)

#> [1] TRUE