Extract source georeferenced catch datasets of WCPFC from Sardara World database

This function extracts the source georeferenced catch datasets stored in the Sardara World database coming from the Western and Central Pacific Ocean Fisheries Commission (WCPFC). The output dataset provides the catch of tuna and tuna-like species in the area managed under the jurisdiction of the WCPFC. Catches are stratified by month, species, gear, fishing mode (i.e. type of school used), area (5° square) and unit of catch (weight or number).

wcpfc_catch_level0(year_tunaatlas)

Arguments

year_tunaatlas
numeric. The year of the datasets to extract. Starts in 2017

Details

The output dataset lists catch of tuna and tuna-like species in the Western Pacific ocean. Catches are stratified by month, species, gear, fishing mode (i.e. type of school used), area (5° square) and unit of catch (weight or number). The dataset is computed using public domain datasets released by the Western and Central Pacific Fisheries Commission (WCPFC). This function merges the primary catch-and-effort datasets released by WCPFC. Output dataset is expressed with WCPFC's coding system.

See also

Other tRFMOs datasets extraction: ccsbt_catch_level0, iattc_catch_level0, iccat_catch_level0, iotc_catch_level0

Examples

# Retrieve WCPFC georeferenced catch data from 2017
wcpfc_catch<-wcpfc_catch_level0(2017)

#> Error in eval(expr, envir, enclos): impossible de trouver la fonction "wcpfc_catch_level0"

head(wcpfc_catch)

#> Error in head(wcpfc_catch): objet 'wcpfc_catch' introuvable